+++ src/main/java/dto/TransactionUpdateDtoIn.java
@@ -1,5 +1,5 @@
 package dto;
 
-public record TransactionUpdateDtoIn(Integer userId, Integer categoryId, Integer transactionAmount, String transactionDate, String transactionType, String transactionDescription, String transactionFrequency) {
+public record TransactionUpdateDtoIn(Integer userId, Integer categoryId, Double transactionAmount, String transactionDate, String transactionType, String transactionDescription, String transactionFrequency) {
 
 }
